requirements.txt
streamlit==1.28.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
Pillow==10.0.1
tensorflow-cpu==2.16.1
protobuf==4.25.0
h5py==3.10.0